[s0463] Fixed some implicit float->int cast warnings.

(based on commit 4c6e61e)

Re-apply small not applied part.

+++ src/game/Player.cpp
@@ -6794,7 +6794,8 @@ void Player::_ApplyItemBonuses(ItemPrototype const *proto,uint8 slot,bool apply)
     uint32 armor = proto->Armor;
     // Add armor bonus from ArmorDamageModifier if > 0
     if (proto->ArmorDamageModifier > 0)
-        armor += proto->ArmorDamageModifier;
+        armor += uint32(proto->ArmorDamageModifier);
+
     if (armor)
         HandleStatModifier(UNIT_MOD_ARMOR, BASE_VALUE, float(armor), apply);
